pyproject.toml
[tool.poetry]
name = "playbacker"
version = "0.1.0"
description = "Live music performance playback"
authors = ["Lev Vereshchagin <mail@vrslev.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
sounddevice = "0.4.4"
SoundFile = "0.10.3-post1"
numpy = "1.22.2"
pydantic = "1.9.0"
PyYAML = "6.0"
textual = "0.1.15"
typer = "0.4.0"
uvloop = "0.16.0"
inquirer = "2.9.1"
soxr = "0.2.7"
textual-inputs = "0.2.5"

[tool.poetry.dev-dependencies]
black = "22.1.0"
pre-commit = "2.17.0"
pytest = "7.0.1"
pytest-cov = "3.0.0"

# https://github.com/python-poetry/poetry-core/pull/182#issuecomment-1005195851
[build-system]
requires = ["poetry-core@https://github.com/python-poetry/poetry-core/archive/325312c016d69189ac93c945ba0c1b69296c5e54.zip"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
playbacker = "playbacker.main:main"

[tool.pyright]
ignore = [".venv"]
reportMissingTypeStubs = false
reportImportCycles = false
reportIncompatibleMethodOverride = false
reportUninitializedInstanceVariable = "warning"
reportImplicitStringConcatenation = "warning"
reportPrivateUsage = false

[tool.black]
target-version = ["py310"]

[tool.isort]
profile = "black"

[tool.coverage.run]
source = ["src/playbacker"]

[tool.coverage.report]
skip_covered = true
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def __str__",
    "@abstractmethod",
]
